Split services generated by volo into several files
