+++ Cargo.toml
@@ -27,9 +27,9 @@ rust-version = "1.80.0"
 #pilota = "0.11"
 #pilota-build = "0.11"
 #pilota-thrift-parser = "0.11"
-pilota = { git = "https://github.com/cloudwego/pilota", branch = "main" }
-pilota-build = { git = "https://github.com/cloudwego/pilota", branch = "main" }
-pilota-thrift-parser = { git = "https://github.com/cloudwego/pilota", branch = "main" }
+pilota = { git = "https://github.com/missingdays/pilota", branch = "fix/context-mode-pub" }
+pilota-build = { git = "https://github.com/missingdays/pilota", branch = "fix/context-mode-pub" }
+pilota-thrift-parser = { git = "https://github.com/missingdays/pilota", branch = "fix/context-mode-pub" }
 
 motore = "0.4"
 # motore = { git = "https://github.com/cloudwego/motore", branch = "main" }
