+++ tests/code-generation-workspace-split/Cargo.toml
@@ -4,8 +4,8 @@ name = "gen"
 test = false
 
 [dependencies.pilota-build]
-branch = "main"
-git = "https://github.com/cloudwego/pilota"
+branch = "fix/context-mode-pub"
+git = "https://github.com/missingdays/pilota"
 
 [dependencies.volo-build]
 workspace = true
@@ -31,8 +31,8 @@ lazy_static = "1"
 serde = "1"
 
 [workspace.dependencies.pilota]
-branch = "main"
-git = "https://github.com/cloudwego/pilota"
+branch = "fix/context-mode-pub"
+git = "https://github.com/missingdays/pilota"
 
 [workspace.dependencies.volo]
 path = "../../volo"
