+++ tests/code-generation-workspace-split/rpc_article/src/article/ArticleService/mod.rs
@@ -1,0 +1,3 @@
+include!("ArticleService.rs");
+include!("enum_ArticleServiceRequestRecv.rs");
+include!("enum_ArticleServiceResponseRecv.rs");
