+++ tests/code-generation-workspace-split/rpc_article/src/article/message_ArticleServiceGetArticleArgsSend.rs
@@ -67,11 +67,7 @@ impl ::pilota::thrift::Message for ArticleServiceGetArticleArgsSend {
             ::std::result::Result::Ok::<_, ::pilota::thrift::ThriftException>(())
         })() {
             if let Some(field_id) = __pilota_decoding_field_id {
-                err.prepend_msg(&format!(
-                    "decode struct `ArticleServiceGetArticleArgsSend` field(#{}) failed, caused \
-                     by: ",
-                    field_id
-                ));
+                err.prepend_msg(&format!("decode struct `ArticleServiceGetArticleArgsSend` field(#{}) failed, caused by: ", field_id));
             }
             return ::std::result::Result::Err(err);
         };
@@ -133,11 +129,7 @@ impl ::pilota::thrift::Message for ArticleServiceGetArticleArgsSend {
             .await
             {
                 if let Some(field_id) = __pilota_decoding_field_id {
-                    err.prepend_msg(&format!(
-                        "decode struct `ArticleServiceGetArticleArgsSend` field(#{}) failed, \
-                         caused by: ",
-                        field_id
-                    ));
+                    err.prepend_msg(&format!("decode struct `ArticleServiceGetArticleArgsSend` field(#{}) failed, caused by: ", field_id));
                 }
                 return ::std::result::Result::Err(err);
             };
