+++ tests/code-generation-workspace-split/rpc_author/src/author/AuthorService/enum_AuthorServiceRequestRecv.rs
@@ -1,0 +1,117 @@
+#[derive(Debug, Clone)]
+pub enum AuthorServiceRequestRecv {
+    GetAuthor(AuthorServiceGetAuthorArgsRecv),
+}
+
+#[derive(Debug, Clone)]
+pub enum AuthorServiceRequestSend {
+    GetAuthor(AuthorServiceGetAuthorArgsSend),
+}
+
+impl ::volo_thrift::EntryMessage for AuthorServiceRequestRecv {
+    fn encode<T: ::pilota::thrift::TOutputProtocol>(
+        &self,
+        __protocol: &mut T,
+    ) -> ::core::result::Result<(), ::pilota::thrift::ThriftException> {
+        match self {
+            Self::GetAuthor(value) => {
+                ::pilota::thrift::Message::encode(value, __protocol).map_err(|err| err.into())
+            }
+        }
+    }
+
+    fn decode<T: ::pilota::thrift::TInputProtocol>(
+        __protocol: &mut T,
+        msg_ident: &::pilota::thrift::TMessageIdentifier,
+    ) -> ::core::result::Result<Self, ::pilota::thrift::ThriftException> {
+        ::std::result::Result::Ok(match &*msg_ident.name {
+            "GetAuthor" => Self::GetAuthor(::pilota::thrift::Message::decode(__protocol)?),
+            _ => {
+                return ::std::result::Result::Err(::pilota::thrift::new_application_exception(
+                    ::pilota::thrift::ApplicationExceptionKind::UNKNOWN_METHOD,
+                    format!("unknown method {}", msg_ident.name),
+                ));
+            }
+        })
+    }
+
+    async fn decode_async<T: ::pilota::thrift::TAsyncInputProtocol>(
+        __protocol: &mut T,
+        msg_ident: &::pilota::thrift::TMessageIdentifier,
+    ) -> ::core::result::Result<Self, ::pilota::thrift::ThriftException> {
+        ::std::result::Result::Ok(match &*msg_ident.name {
+            "GetAuthor" => Self::GetAuthor(
+                <AuthorServiceGetAuthorArgsRecv as ::pilota::thrift::Message>::decode_async(
+                    __protocol,
+                )
+                .await?,
+            ),
+            _ => {
+                return ::std::result::Result::Err(::pilota::thrift::new_application_exception(
+                    ::pilota::thrift::ApplicationExceptionKind::UNKNOWN_METHOD,
+                    format!("unknown method {}", msg_ident.name),
+                ));
+            }
+        })
+    }
+
+    fn size<T: ::pilota::thrift::TLengthProtocol>(&self, __protocol: &mut T) -> usize {
+        match self {
+            Self::GetAuthor(value) => ::volo_thrift::Message::size(value, __protocol),
+        }
+    }
+}
+
+impl ::volo_thrift::EntryMessage for AuthorServiceRequestSend {
+    fn encode<T: ::pilota::thrift::TOutputProtocol>(
+        &self,
+        __protocol: &mut T,
+    ) -> ::core::result::Result<(), ::pilota::thrift::ThriftException> {
+        match self {
+            Self::GetAuthor(value) => {
+                ::pilota::thrift::Message::encode(value, __protocol).map_err(|err| err.into())
+            }
+        }
+    }
+
+    fn decode<T: ::pilota::thrift::TInputProtocol>(
+        __protocol: &mut T,
+        msg_ident: &::pilota::thrift::TMessageIdentifier,
+    ) -> ::core::result::Result<Self, ::pilota::thrift::ThriftException> {
+        ::std::result::Result::Ok(match &*msg_ident.name {
+            "GetAuthor" => Self::GetAuthor(::pilota::thrift::Message::decode(__protocol)?),
+            _ => {
+                return ::std::result::Result::Err(::pilota::thrift::new_application_exception(
+                    ::pilota::thrift::ApplicationExceptionKind::UNKNOWN_METHOD,
+                    format!("unknown method {}", msg_ident.name),
+                ));
+            }
+        })
+    }
+
+    async fn decode_async<T: ::pilota::thrift::TAsyncInputProtocol>(
+        __protocol: &mut T,
+        msg_ident: &::pilota::thrift::TMessageIdentifier,
+    ) -> ::core::result::Result<Self, ::pilota::thrift::ThriftException> {
+        ::std::result::Result::Ok(match &*msg_ident.name {
+            "GetAuthor" => Self::GetAuthor(
+                <AuthorServiceGetAuthorArgsSend as ::pilota::thrift::Message>::decode_async(
+                    __protocol,
+                )
+                .await?,
+            ),
+            _ => {
+                return ::std::result::Result::Err(::pilota::thrift::new_application_exception(
+                    ::pilota::thrift::ApplicationExceptionKind::UNKNOWN_METHOD,
+                    format!("unknown method {}", msg_ident.name),
+                ));
+            }
+        })
+    }
+
+    fn size<T: ::pilota::thrift::TLengthProtocol>(&self, __protocol: &mut T) -> usize {
+        match self {
+            Self::GetAuthor(value) => ::volo_thrift::Message::size(value, __protocol),
+        }
+    }
+}
