+++ tests/code-generation-workspace-split/rpc_author/src/author/AuthorService/mod.rs
@@ -1,0 +1,3 @@
+include!("AuthorService.rs");
+include!("enum_AuthorServiceRequestRecv.rs");
+include!("enum_AuthorServiceResponseRecv.rs");
