+++ tests/code-generation-workspace-split/rpc_image/src/article/image/ImageService/mod.rs
@@ -1,0 +1,3 @@
+include!("ImageService.rs");
+include!("enum_ImageServiceRequestRecv.rs");
+include!("enum_ImageServiceResponseRecv.rs");
