+++ tests/code-generation-workspace-split/rpc_image/src/article/image/message_ImageServiceGetImageArgsRecv.rs
@@ -132,11 +132,7 @@ impl ::pilota::thrift::Message for ImageServiceGetImageArgsRecv {
             .await
             {
                 if let Some(field_id) = __pilota_decoding_field_id {
-                    err.prepend_msg(&format!(
-                        "decode struct `ImageServiceGetImageArgsRecv` field(#{}) failed, caused \
-                         by: ",
-                        field_id
-                    ));
+                    err.prepend_msg(&format!("decode struct `ImageServiceGetImageArgsRecv` field(#{}) failed, caused by: ", field_id));
                 }
                 return ::std::result::Result::Err(err);
             };
